+++ android/src/main/java/io/parity/signer/ui/navigationselectors/SignerSelectorsLegacy.kt
@@ -11,6 +11,7 @@ import io.parity.signer.models.*
 import io.parity.signer.screens.*
 import io.parity.signer.screens.keysets.*
 import io.parity.signer.screens.keysets.details.SeedMenu
+import io.parity.signer.screens.keysets.details.backup.SeedBackup
 import io.parity.signer.uniffi.Action
 import io.parity.signer.uniffi.AlertData
 import io.parity.signer.uniffi.ModalData
